pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miibrowser"
version = "1.0.0"
authors = [
    { name = "MiiBrowser Team", email = "team@miibrowser.example.com" }
]
description = "A simple web browser with DuckDuckGo search capabilities"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["browser", "duckduckgo", "search", "tkinter", "gui", "chromium", "webview"]
dependencies = [
    "requests>=2.31.0",
    "pywebview>=6.0.0",
    "Pillow>=9.0.0",
    "tinycss2>=1.2.0",
    "esprima>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
legacy = [
    "tkinterweb>=3.18.0",
    "CairoSVG>=2.5.0",
    "tkinterweb-tkhtml-extras>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/MiiBrowser"
"Bug Tracker" = "https://github.com/yourusername/MiiBrowser/issues"

[project.scripts]
miibrowser = "miibrowser.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["src/miibrowser"]
branch = true

[tool.coverage.report]
show_missing = true
precision = 2
